requirements.txt
streamlit
torch
transformers
python-docx
python-pptx
PyPDF2